+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- More detailed logging of Type decisions when extracting to database
+
 ## [4.2.4] - 2021-02-05
 
 ### Added